+++ src/TestSuite/MockTrait.php
@@ -17,6 +17,7 @@ namespace MeTools\TestSuite;
 
 use Cake\Controller\ComponentRegistry;
 use Cake\Controller\Controller;
+use Cake\View\Helper;
 use Cake\View\View;
 use PHPUnit\Framework\TestCase;
 
@@ -91,13 +92,14 @@ trait MockTrait
      * Mocks an helper
      * @param class-string<\Cake\View\Helper> $className Helper class name
      * @param array<string>|null $methods The list of methods to mock
+     * @param \Cake\View\View|null $View A `View` instance
      * @return \Cake\View\Helper&\PHPUnit\Framework\MockObject\MockObject
      */
-    protected function getMockForHelper(string $className, ?array $methods = []): object
+    protected function getMockForHelper(string $className, ?array $methods = [], ?View $View = null): Helper
     {
         return $this->getMockBuilder($className)
             ->setMethods($methods)
-            ->setConstructorArgs([new View()])
+            ->setConstructorArgs([$View ?: new View()])
             ->getMock();
     }