+++ src/TestSuite/TestCase.php
@@ -39,18 +39,18 @@ abstract class TestCase extends CakeTestCase
      * Called before every test method
      * @return void
      */
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
-        $this->loadPlugins(Configure::read('pluginsToLoad') ?: ['MeTools' => []]);
+        $this->loadPlugins(Configure::read('pluginsToLoad', ['MeTools' => []]));
     }
 
     /**
      * Called after every test method
      * @return void
      */
-    public function tearDown(): void
+    protected function tearDown(): void
     {
         parent::tearDown();