+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # 2.x branch
 ## 2.21 branch
+### 2.21.1
+* added a theme for Bake. See the `README` file;
+* added `IntegrationTestTrait::getStatusCode()` method;
+* fixed a little bug for `HtmlHelper::link()`;
+* added `BootstrapFormHelper`. For now this class is temporary, but in the future
+    it will replace `FormHelper`, assuming the same name. Some methods of the
+    `FormHelperTest` test class have been marked as deprecated, to indicate that
+    code is already covered by the new code;
+* `datepicker()`, `datetimepicker()` and `timepicker()` methods provided by
+    `FormHelper` are now deprecated and they will be removed in a later version.
+    Use instead the normal `control()` method, which will generate
+    `date`/`datetime-local`/`time` inputs, recognized by the browser;
+* `datepicker()`, `datetimepicker()` and `timepicker()` methods provided by
+    LibraryHelper` are now deprecated and they will be removed in a later version;
+* little fixes for `FormHelper`, to make the output more consistent with what
+    Bootstrap requires;
+* `MockTrait::getMockForHelper()` method has `$view` argument;
+* updated some css file, to make the output more consistent with what Bootstrap
+    requires;
+* added some i18n constants (`config/i18n_constants.php`), in order to be used universally;
+* fixed the `flash` template element;
+* requires at least CakePHP 4.2.
+
 ### 2.21.0
 * numerous code adjustments for improvement and adaptation to PHP 7.4 new features;
 * `OptionsParser` now uses the `$defaults` array property to store default values.